+++ src/Ext/House/Hooks.cpp
@@ -258,6 +258,27 @@ DEFINE_HOOK(0x7015C9, TechnoClass_Captured_UpdateTracking, 0x6)
 		vec.erase(std::remove(vec.begin(), vec.end(), pThis), vec.end());
 	}
 
+	if (pExt->TypeExtData->UniqueTechno)
+	{
+		const auto pOldOwner = pThis->Owner;
+
+		if (pOldOwner->IsControlledByCurrentPlayer())
+		{
+			if (!pNewOwner->IsControlledByCurrentPlayer())
+			{
+				auto& vec = ScenarioExt::Global()->OwnedUniqueTechnos;
+				vec.erase(std::remove(vec.begin(), vec.end(), pExt), vec.end());
+			}
+		}
+		else if (pNewOwner->IsControlledByCurrentPlayer())
+		{
+			auto& vec = ScenarioExt::Global()->OwnedUniqueTechnos;
+
+			if (std::find(vec.begin(), vec.end(), pExt) == vec.end())
+				vec.push_back(pExt);
+		}
+	}
+
 	if (auto pMe = generic_cast<FootClass*>(pThis))
 	{
 		bool I_am_human = pThis->Owner->IsControlledByHuman();